+++ packages/vite/src/plugins/index.ts
@@ -2,7 +2,9 @@ import type { Options } from "@srvkit/common/@types/options/default";
 import type { ResolvedOptions } from "@srvkit/common/@types/options/resolved";
 import type { Plugin } from "vite";
 
+import { defineEnv } from "@srvkit/common/functions/env/define";
 import { resolveOptions } from "@srvkit/common/functions/options/resolve";
+import { replacePlugin } from "rolldown/plugins";
 
 import { buildPlugin } from "#/plugins/build";
 import { copyPlugin } from "#/plugins/copy";
@@ -12,6 +14,19 @@ const plugin = (options?: Options): Plugin[] => {
     const opts: ResolvedOptions = resolveOptions(options);
 
     return [
+        replacePlugin(
+            defineEnv({
+                fallback: "production",
+            }),
+            {
+                delimiters: [
+                    "\\b",
+                    "(?![A-Za-z0-9_$])",
+                ],
+                preventAssignment: true,
+                sourcemap: true,
+            },
+        ),
         devPlugin({
             ...opts,
         }),